+++ src/Console/PublishCommand.php
@@ -1,177 +1,37 @@
 <?php namespace Orchestra\Studio\Console;
 
 use Illuminate\Console\Command;
-use League\Flysystem\MountManager;
 use Illuminate\Filesystem\Filesystem;
-use League\Flysystem\Filesystem as Flysystem;
+use Orchestra\Studio\Traits\PublishFilesTrait;
 use Symfony\Component\Console\Input\InputOption;
-use League\Flysystem\Adapter\Local as LocalAdapter;
 
 abstract class PublishCommand extends Command
 {
-    /**
-     * The filesystem instance.
-     *
-     * @var \Illuminate\Filesystem\Filesystem
-     */
-    protected $files;
-
-    /**
-     * Create a new command instance.
-     *
-     * @param  \Illuminate\Filesystem\Filesystem  $files
-     *
-     * @return void
-     */
-    public function __construct(Filesystem $files)
-    {
-        parent::__construct();
-
-        $this->files = $files;
-    }
+    use PublishFilesTrait;
 
     /**
      * Execute the console command.
      *
      * @return void
      */
-    public function handle()
+    public function handle(Filesystem $filesystem)
     {
-        $this->publishFiles();
-
-        if ($this->option('routes') || $this->confirm('Do you wish to append routes? [y|N]')) {
-            $this->appendRoutes();
-        }
+        $this->publishing($filesystem, $this->option('force'));
 
         $this->info('Publishing complete!');
     }
 
     /**
-     * Publish files.
+     * Publishing command.
+     *
+     * @param  \Illuminate\Filesystem\Filesystem  $filesystem
+     * @param  bool    $force
      *
      * @return void
      */
-    protected function publishFiles()
+    protected function publishing(Filesystem $filesystem, $force = false)
     {
-        $paths = (array) $this->publishes();
-
-        foreach ($paths as $from => $to) {
-            if ($this->files->isFile($from)) {
-                $this->publishFile($from, $to);
-            } elseif ($this->files->isDirectory($from)) {
-                $this->publishDirectory($from, $to);
-            } else {
-                $this->error("Can't locate path: <{$from}>");
-            }
-        }
-    }
-
-    /**
-     * Appends routes.
-     *
-     * @return void
-     */
-    protected function appendRoutes()
-    {
-        $file      = $this->getRoutesFile();
-        $routeFile = app_path('Http/routes.php');
-
-        if ($this->files->exists($routeFile) && $this->files->exists($file)) {
-            $this->files->append($routeFile, $this->buildFile($file));
-            $this->line('<info>Append routes from</info> <comment>['.$file.']</comment>');
-        }
-    }
-
-    /**
-     * Publish the file to the given path.
-     *
-     * @param  string  $from
-     * @param  string  $to
-     *
-     * @return void
-     */
-    protected function publishFile($from, $to)
-    {
-        if ($this->files->exists($to) && ! $this->option('force')) {
-            return;
-        }
-
-        $content = $this->buildFile($from);
-
-        $this->createParentDirectory(dirname($to));
-        $this->files->put($to, $content);
-
-        $this->status($from, $to, 'File');
-    }
-
-    /**
-     * Publish the directory to the given directory.
-     *
-     * @param  string  $from
-     * @param  string  $to
-     *
-     * @return void
-     */
-    protected function publishDirectory($from, $to)
-    {
-        $manager = new MountManager([
-            'from' => new Flysystem(new LocalAdapter($from)),
-            'to'   => new Flysystem(new LocalAdapter($to)),
-        ]);
-
-        foreach ($manager->listContents('from://', true) as $file) {
-            if ($file['type'] === 'file' && (! $manager->has('to://'.$file['path']) || $this->option('force'))) {
-                $content = $this->replaceNamespace($manager->read('from://'.$file['path']));
-
-                $manager->put('to://'.$file['path'], $content);
-            }
-        }
-
-        $this->status($from, $to, 'Directory');
-    }
-    /**
-     * Create the directory to house the published files if needed.
-     *
-     * @param  string  $directory
-     *
-     * @return void
-     */
-    protected function createParentDirectory($directory)
-    {
-        if (! $this->files->isDirectory($directory)) {
-            $this->files->makeDirectory($directory, 0755, true);
-        }
-    }
-
-    /**
-     * Write a status message to the console.
-     *
-     * @param  string  $from
-     * @param  string  $to
-     * @param  string  $type
-     *
-     * @return void
-     */
-    protected function status($from, $to, $type)
-    {
-        $from = str_replace(base_path(), '', realpath($from));
-        $to   = str_replace(base_path(), '', realpath($to));
-
-        $this->line('<info>Copied '.$type.'</info> <comment>['.$from.']</comment> <info>To</info> <comment>['.$to.']</comment>');
-    }
-
-    /**
-     * Build the file with the given path.
-     *
-     * @param  string  $path
-     *
-     * @return string
-     */
-    protected function buildFile($path)
-    {
-        $stub = $this->files->get($path);
-
-        return $this->replaceNamespace($stub);
+        $this->publishFiles($filesystem, (array) $this->publishes(), $force);
     }
 
     /**
@@ -218,11 +78,4 @@ abstract class PublishCommand extends Command
      * @return array
      */
     abstract protected function publishes();
-
-    /**
-     * Get routes file.
-     *
-     * @return array
-     */
-    abstract protected function getRoutesFile();
 }